+++ sql/prolog_queries/entrega/produtividade/func_get_produtividade_colaborador.sql
@@ -27,15 +27,15 @@ create or replace function func_get_produtividade_colaborador(f_mes integer, f_a
                 kmprevistoroad             real,
                 kmsai                      integer,
                 kmentr                     integer,
-                tempoprevistoroad          integer,
+                tempoprevistoroad          bigint,
                 hrsai                      timestamp without time zone,
                 hrentr                     timestamp without time zone,
                 tempo_rota                 bigint,
-                tempointerno               integer,
+                tempointerno               bigint,
                 hrmatinal                  time without time zone,
                 apontamentos_ok            bigint,
                 total_tracking             bigint,
-                tempo_largada              integer,
+                tempo_largada              bigint,
                 meta_tracking              real,
                 meta_tempo_rota_mapas      real,
                 meta_caixa_viagem          real,
@@ -48,24 +48,24 @@ create or replace function func_get_produtividade_colaborador(f_mes integer, f_a
                 meta_raio_tracking         real,
                 meta_tempo_interno_mapas   real,
                 meta_tempo_largada_mapas   real,
-                meta_tempo_rota_horas      integer,
-                meta_tempo_interno_horas   integer,
-                meta_tempo_largada_horas   integer,
-                meta_jornada_liquida_horas integer,
+                meta_tempo_rota_horas      bigint,
+                meta_tempo_interno_horas   bigint,
+                meta_tempo_largada_horas   bigint,
+                meta_jornada_liquida_horas bigint,
+                rm_numero_viagens          smallint,
                 valor_rota                 real,
                 valor_recarga              real,
                 valor_diferenca_eld        double precision,
                 valor_as                   real,
-                valor                      double precision,
-                rm_numero_viagens          smallint
+                valor                      double precision
             )
     language sql
 as
 $$
 select *
-from view_produtividade_extrato
-where data between func_get_data_inicio_produtividade(f_ano, f_mes, f_cpf, null) and
+from view_produtividade_extrato_com_total vpe
+where vpe.data between func_get_data_inicio_produtividade(f_ano, f_mes, f_cpf, null) and
     func_get_data_fim_produtividade(f_ano, f_mes, f_cpf, null)
-  and cpf = f_cpf
-order by data asc
+  and vpe.cpf = f_cpf
+order by vpe.data
 $$;